move ingest CT data
